backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "fastapi (>=0.136.0,<0.137.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# === Общие настройки окружения ===
target-version = "py314"       # В 2026 минимальная версия — Python 3.12 или 3.13
line-length = 79              # Оптимально для современных широких мониторов и типизированного кода
indent-width = 4

# === Глобальное исключение директорий ===
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "migrations",               # Django / Alembic
    "node_modules",
    "*.ipynb_checkpoints",
]

# === Настройки линтинга (Правила) ===
[tool.ruff.lint]
ignore = [
  'B006', # Learn Python yo!
  'C408',  # Unnecessary dict/list call
  'COM819',  # Trailing commas is da bomb
  'E731',  # Do not assign to lambda
  'ERA001', 'T201', # comment code, prints are lax during development
  'G004', # Logging format string should not use f-string
  'RET503', # No explicit return None (all the other rules are about removing unnecessary things like this)
  'RUF012', # Mutable class attributes should be annotated with typing.ClassVar
  'S311', # Standard pseudo-random generators are not suitable for security/cryptographic purposes
  'SIM108', # don't like ternary operator
  'SIM300', # Yoda is wiser than you!
  'TRY003',  # Avoid long messages outside of exception class
  #'TRY004',  # Use TypeError instead of ValueError
  'TRY301',  # Abstract raise garbage
]

select = [
  'A',  # flake8-builtins
  'ASYNC', # https://docs.astral.sh/ruff/rules/#flake8-async-async
  'B',  # flake8-bugbear
  'BLE', # flake8-blind-except
  'C4',  # unnecessary comprehensions, map()
  'COM',  # flake8-commas
  'DTZ', # flake8-datetimez
  'E',  # pycodestyle
  'ERA',  # No commented out code
  'EXE', # flake8-executable
  'F',  # pyflakes
  'FLY',  # flynt
  'G',  # flake8-logging-format
  'I',  # isort
  'ICN', # https://github.com/joaopalmeiro/flake8-import-conventions
  'ISC', # https://pypi.org/project/flake8-implicit-str-concat/
  'LOG', # flake8-logging
  'PERF', # perflint
  'PIE', # https://pypi.org/project/flake8-pie/
  'PLC',  # Pylint conventions
  'PLE',  # Pylint error
  'PLW',  # Pylint warnings
  'PT',  # https://pypi.org/project/flake8-pytest-style/
  'PTH',  # flake8 use pathlib
  'RET', # https://pypi.org/project/flake8-return/
  'RUF',  # Ruff rules
  'S',  # https://docs.astral.sh/ruff/rules/#flake8-bandit-s
  'SIM',  # https://pypi.org/project/flake8-simplify/
  'T',  # flake8-debugger
  'TRY',  # tryceratops
  'UP',  # pyupgrade
  'W',  # pycodestyle
  #'ARG',  # flake8 unused arguments (not really helpful, unused func args are common and ok)
  #'D',  # pydocstyle (too much)
  #'N',  # pep8-naming (too opinionated)
  #'NPY', # numpy
  #'PD', # pandas
  #'PL',  # Full Pylint (too much)
  #'PLR',  # Pylint refactor (too much/too opinionated)
]

[tool.ruff.format]
quote-style = "double"          # "Черный" стиль (PEP 8 рекомендует двойные)
indent-style = "space"
skip-magic-trailing-comma = false # Всегда добавляем запятую в конце многострочных списков
line-ending = "lf"
docstring-code-format = true    # Форматировать примеры кода в docstring
preview = true     